fix keyvalues encode crash
Accomplished in exchange for my sanity.

Effectively, the bug is that m_pSub can be null, so when it does
`dat->m_pSub->WriteAsBinary( buffer );`, it would be running
WriteAsBinary with `this` set to null. So, if we detect a null m_pSub,
we write a empty keyvalues structure manually by writing the ending
marker directly. The previous "fix" for this accidentally did exactly
this, just in a more roundabout way.

Builds under scout (including official builds) don't have this issue,
due to the version of GCC in scout not being new enough to optimise
this.